COH-502: Bridging Jersey's JavaUtil Logging to slf4j
By default, Jersey logs out using java-util-logging. This created
annoying log messages in STDERR and results in jersey logs missing the
main log.

This commit installs the JUL-slf4j bridge in the JerseyServletModule
using a static initialiser. Any existing JavaUtil loggers are
deregistered first to remove the stderr logging.

In order to test this, I needed to start Jersey, so I added some
acceptance tests to test the Filter chains and the exception handling.
As part of this testing, a default retry-after value was added to the
ServiceUnavailableException (as it's not valid to throw this without a
retry header).

Change-Id: I35c5bba56dba9d680b680564ff83f65213079454