Provide some basic impls and modules for required classes
Added a permissive Authenticator impl and a module to bind it.
Also, a module that injects a ServerInfo object that reads its
identifier value from a system property. Updated the api tests to
use these new modules so we could lose the cludgy StubModule
implementations in the tests.

Finally, updated metrics version to 2.0.0-BETA18 as this is
available in maven central, so we don't need to specify
coda's repo anymore.

Change-Id: Ia6d09961c118513ee8ee10592f903f09602c740e